data_management: fix deadlock on failed syncs
The sync was kept in an errnoeous state after it is failing to allocate
enough space in the storage, which lead to deadlocks when dropping the
database object for example.

This commit returns the MID into it's original state and simply will
lead future sync to fail the same way as the first one.
To fix the problem from a user perspective the space accounting should
be expanded to fail early on inserts which exceeded the recorded
capacity.