dataset: Use messages for migration
This currently has the down side that the data is only migrated once it
reaches the leafs which depending on the structure of the tree might
take some time. This can be solved by implementing a force flush on
migrations so that all relevant data reaches the leafs on insert.
Though this may require in the end a new message type, which extends the
upsert one.