object: use generic checksum instead of xxhash
This is from the artifacts in which we've tested the way of comparing in the
unpack function or potentially later to see if data is actually valid. It is
unclear if we actucally need this. The hash from the initial read is anyway
difficult to verify properly later as we throw this buffer away.